requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0  # For scheduling functionality
openai>=1.0.0
selenium>=4.0.0  # For web automation if needed 